kc/requirements.txt
kucoin-python==1.0.11
